Tools 2.0.0
This has significantly changed and (I think) improved the interfaces
to require-module.

require-module now passes arguments to wrapper via a
:wrapper-arguments option rather than handing any additional keywords
to them.  This means it's a lot easier to detect bad arglists, at the
small cost of making passing arguments to wrappers a little harder.
Since I've only ever done that to test it works, that's a small cost.

require-modules now takes module specifications which are either
module names or a cons of (module-name . args-to-require-module).  So
you can require two modules with only one being verbose.

requires is a new function which is a nospread version of
require-modules with a fallback to require.

needs is a new macro which is requires wrapped in a suitable eval-when
to make it happen at compile-time as well (this was previously a
common pattern).

provides is to provide-module as requires is to require-module.